requirements.txt
google-genai>=1.51.0
Pillow
numpy
opencv-python-headless
